feat: implements scroll visualization

+++ source/Age/Elements/Element.LayoutCommand.cs
@@ -5,15 +5,15 @@ namespace Age.Elements;
 public abstract partial class Element
 {
     [KeyedListKey]
-    internal enum LayoutCommand : byte
+    internal enum LayoutCommand : ushort
     {
-        None     = 0,
-        Box      = 1 << 0,
-        Image    = 1 << 1,
-        ScrollX  = 1 << 2,
-        ScrollY  = 1 << 3,
-        Layer1   = 1 << 4,
-        Layer2   = 1 << 5,
-        Layer3   = 1 << 6,
+        None    = 0,
+        Box     = 1 << 0,
+        Image   = 1 << 1,
+        ScrollX = 1 << 2,
+        ScrollY = 1 << 3,
+        Layer1  = 1 << 4,
+        Layer2  = 1 << 5,
+        Layer3  = 1 << 6,
     }
 }
